+++ src/nveditorlib/text_box.py
@@ -10,23 +10,102 @@ from tkinter import scrolledtext
 
 
 class TextBox(scrolledtext.ScrolledText):
-    """If a more sophisticated text box is needed, create it here."""
+    """A text editor widget for yWriter raw markup.
+    
+    Public methods:
+    get_text -- Return the whole text from the editor box.
+    set_text(text) -- Put text into the editor box and clear the undo/redo stack.
+    count_words -- Return the word count.
+    italic -- Make the selection italic, or begin with italic input.
+    bold -- Make the selection bold, or begin with bold input.
+    plain -- Remove formatting from the selection.
+    """
+    _YW_TAGS = ('i', 'b')
+    # Supported tags.
 
     def get_text(self):
+        """Return the whole text from the editor box."""
         text = self.get('1.0', tk.END).strip(' \n')
-        # convert text to yWriter markup, if applicable.
         return text
 
     def set_text(self, text):
-        # convert text from yWriter markup, if applicable.
+        """Put text into the editor box and clear the undo/redo stack."""
         self.insert(tk.END, text)
         self.edit_reset()
         # this is to prevent the user from clearing the box with Ctrl-Z
 
     def count_words(self):
+        """Return the word count."""
         text = re.sub('--|—|–|…', ' ', self.get('1.0', tk.END))
         # Make dashes separate words
         text = re.sub('\[.+?\]|\/\*.+?\*\/|\.|\,|-', '', text)
         # Remove comments and yWriter raw markup for word count; make hyphens join words
         wordList = text.split()
         return len(wordList)
+
+    def italic(self, event=None):
+        """Make the selection italic, or begin with italic input."""
+        self._set_format(tag='i')
+
+    def bold(self, event=None):
+        """Make the selection bold, or begin with bold input."""
+        self._set_format(tag='b')
+
+    def plain(self, event=None):
+        """Remove formatting from the selection."""
+        self._set_format()
+
+    def _set_format(self, event=None, tag=''):
+        """Insert an opening/closing pair of yWriter markup tags."""
+        if tag:
+            # Toggle format as specified by tag.
+            if self.tag_ranges(tk.SEL):
+                text = self.get(tk.SEL_FIRST, tk.SEL_LAST)
+                if text.startswith(f'[{tag}]'):
+                    if text.endswith(f'[/{tag}]'):
+                        # The selection is already formatted: Remove markup.
+                        text = self._remove_format(text, tag)
+                        self._replace_selected(text)
+                        return
+
+                # Format the selection: Add markup.
+                text = self._remove_format(text, tag)
+                # to make sure that there is no nested markup of the same type
+                self._replace_selected(f'[{tag}]{text}[/{tag}]')
+            else:
+                # Add markup to the cursor position.
+                self.insert(tk.INSERT, f'[{tag}]')
+                endTag = f'[/{tag}]'
+                self.insert(tk.INSERT, endTag)
+                self.mark_set(tk.INSERT, f'{tk.INSERT}-{len(endTag)}c')
+        elif self.tag_ranges(tk.SEL):
+            # Remove all markup from the selection.
+            text = self.get(tk.SEL_FIRST, tk.SEL_LAST)
+            for tag in self._YW_TAGS:
+                text = self._remove_format(text, tag)
+            self._replace_selected(text)
+
+    def _replace_selected(self, text):
+        """Replace the selected passage by text; keep the selection."""
+        self.mark_set(tk.INSERT, tk.SEL_FIRST)
+        self.delete(tk.SEL_FIRST, tk.SEL_LAST)
+        selFirst = self.index(tk.INSERT)
+        self.insert(tk.INSERT, text)
+        selLast = self.index(tk.INSERT)
+        self.tag_add(tk.SEL, selFirst, selLast)
+
+    def _remove_format(self, text, tag):
+        """Return text without opening/closing markup, if any."""
+        if tag in self._YW_TAGS:
+            finished = False
+            while not finished:
+                start = text.find(f'[{tag}]')
+                if start >= 0:
+                    end = text.find(f'[/{tag}]')
+                    if  start < end:
+                        text = f'{text[:start]}{text[start + 3:end]}{text[end + 4:]}'
+                    else:
+                        finished = True
+                else:
+                    finished = True
+            return text